chore: reuse existing instances
**before:**
When a user changes the type of a mesh, camera or light.
the instance would be destroyed and it would be attempted to create
new one.

Therefore, when the type is incomplete (while typing), the instance
"disappears" (for the viewer).

**now**
If a previous instance exists but the "class" of the object cannot be
determined, that previous instance will be reused.